Update the manage lists page and User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -51,6 +51,10 @@ class ListingController extends Controller
     }
     public function update(Listing $listing)
     {
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, "Unauthorized Action");
+        }
         $listsUpdateData = request()->validate([
             'company' => ['required'],
             'title' => ['required'],
@@ -68,6 +72,9 @@ class ListingController extends Controller
     }
     public function destroy(Listing $listing)
     {
+        if($listing->user_id != auth()->id()){
+            abort(403, "Unauthorized Action");
+        }
         $listing->delete();
         return redirect('/')->with('flashMessage', "List Deleted Sucessfully!");
     }
